De Bruijn Graphs
Given an arbitrary collection of k-mers (or an integer and string), every edge is labeled by a k-mer (or substring). The starting and ending nodes of an edge (or substring) is labeled by the prefix and suffix of the k-mer (or substring) labeling that edge. We then define the de Bruijn graph by pairing identically labeled nodes, which yields the De Bruijn graph.in the form of an adjacency list or