Overlap Graph
Given a collection of k-mers Patterns, create a graph with a node for each k-mer in the collection and connect k-mers by a directed edge if the suffix is equal to the prefix of another kmer. The resulting graph is called the overlap graph presented in the form of an adjacency list.